Implement bulkcmd, used to send commands to U-Boot
See [here][u-boot-impl] for the gadget-side implementation of this in
Amlogic's U-Boot. Even though I don't believe this is implemented in
ROMs, it belongs in this tool because the U-Boot implementation speaks
basically the same protocol as the ROMs do.

Also add an extremely simple script to send its argument as a command to
a connected device.

[u-boot-impl]: https://github.com/khadas/u-boot/blob/4c6694f6a8c250fcf88b4cdd1c76587f2b04d21a/drivers/usb/gadget/v2_burning/v2_usb_tool/usb_pcd.c#L703-L708