Add NUL terminator to tplCommand() parameter
The U-Boot implementation of this treats the argument as a string but
does not add a terminator or clear the buffer between invocations. The
result of this is that (apart from potentially leaking random stack
data), commands can get clobbered by pieces of previous, longer
invocations that are still in memory. Add a terminator on our end to
avoid this.