runKernel: use writeLargeMemory() for device tree
Both the kernel and ramdisk are currently transferred with
writeLargeMemory(), but the device tree uses writeMemory(). This is
leading to timeouts on my system, trying to boot a S905Y2 (Meson G12A)
with the device tree that currently exists for it in mainline. Switch to
writeLargeMemory() like the others to avoid the issue.