runKernel: escape command line by default
The kernel command line is set using a U-Boot command which is subject
to parsing by the U-Boot shell. Automatically add quoting and escape
sequences to prevent this parsing from mangling the string.

For cases where U-Boot parsing is desired (for example, variable
substitution), add a new `-n` flag to disable the new behavior.