Prefix libretech board names with the SoC to separate form factor code
The naming convention of the Libre Computer boards is described in the
link below:

https://hub.libre.computer/t/libre-computer-board-naming-scheme-and-conventions/100

PLAT_VENDOR-PLAT_FAMILY-FORM_FACTOR-VARIANT-VERSION

The naming convention of libretech followed by the form factor causes
conflicts when the SoC is different but the form factor is the same.

This patch prefixes the SoC/PLAT_FAMILY to the form factor to create
unique names for Amlogic-based boards.