pyamlboot: change amlogic SoC detection procedure
Device detection procedure was modified to wait the response
with amlogic identity. In case when device is starting from
DDR copy of uboot the device appears on the bus earlier than
it enters to flashing mode. When we send requests for flash
during switch interval, we get wrong responses, that's why
we need to wait until device will be fully initialized.

Signed-off-by: Vladimir Mitrofanov <vvmitrofanov@salutedevices.com>
Change-Id: I88e3f5ab60179d9cd41dc72f63f5b8856b30f64b